# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Test",
    "description": "",
    "version": "1.0.0"
  },
  "host": "",
  "basePath": "/",
  "tags": [
    {
      "name": "API",
      "description": "API endpoints"
    },
    {
      "name": "Pages",
      "description": "HTML pages"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/signup": {
      "post": {
        "tags": [
          "API"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "login": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "password2": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "get": {
        "tags": [
          "Pages"
        ],
        "description": "",
        "produces": [
          "text/html"
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": [
          "API"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "login": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Pages"
        ],
        "description": "",
        "produces": [
          "text/html"
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": [
          "API"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Pages"
        ],
        "description": "",
        "produces": [
          "text/html"
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/get-user-info": {
      "get": {
        "tags": [
          "API"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/home": {
      "get": {
        "tags": [
          "Pages"
        ],
        "description": "",
        "produces": [
          "text/html"
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "Pages"
        ],
        "description": "",
        "produces": [
          "text/html"
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    }
  }
}